+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Added a quick "God check" to login objects.
 - Added the ability to get every single token, when in "God Mode."
 - Made sure that the login_id field for Postgres Security is "unique."
+- Added a function to test tokens.
+- Added a method to check the access class of rows quickly.
+- Added methods to validate visibility of rows.
 
 **1.0.0.2035** *(July 6, 2018)*